requirements.txt
daphne
Django
django-ckeditor
django-cors-headers
django-environ
django-js-asset
gunicorn
hyperlink
channels
djangorestframework
Pillow
django-storages
psycopg2
pandas
sqlalchemy
pymysql
mysqlclient
redis